refactor: change common.js to createElement.js
