+++ scripts/components/hint.js
@@ -1,4 +1,4 @@
-import { createElement } from '../common.js';
+import createElement from '../createElement.js';
 
 const subtitileText = 'Alt + Control to change Language';
 const hintSubtitile = createElement('h3', ['text-area__subtitle', 'hint']);
